requirements.txt
blurhash==1.1.4
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
decorator==5.1.1
Flask==2.2.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Mastodon.py==1.8.0
numpy==1.24.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-magic==0.4.27
requests==2.28.2
scipy==1.9.3
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.14.0
